Initial commit of mongosh
Summary:
mongosh is a new REPL for managing MongoDB instances and querying with them.
Because the old MongoDB shell, a.k.a. `mongo`, is going to be eventually fully deprecated and imposes more build time in the mongodb package, this might need to be included.

Signed-off-by: Martin Reboredo <yakoyoku@gmail.com>

Test Plan: Performed some queries as described in the [mongosh guide](https://docs.mongodb.com/mongodb-shell/run-commands/).

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D11619